+++ commands/charity.js
@@ -1,0 +1,4 @@
+module.exports = function charity({ chat }) {
+  const msg = 'All proceeds go to the San Francisco-Marin Food Bank. You can donate directly here: https://us-p2p.netdonor.net/1803/general/61375/cgbuen'
+  return msg
+}
