+++ commands/commands.js
@@ -1,0 +1,4 @@
+module.exports = function commands({ chat }) {
+  const msg = '!commands / !fc / !discord / !controls / !so [user] / !song / !charity / !lurk / !uptime / !chrissucks / !rank, more info in the channel note panels below'
+  return msg
+}
