+++ commands/controls.js
@@ -1,0 +1,4 @@
+module.exports = function controls({ chat }) {
+  const msg = 'pro controller, motion 4.0, R stick 0'
+  return msg
+}
