+++ commands/discord.js
@@ -1,0 +1,6 @@
+const { DISCORD } = require('../vars')
+
+module.exports = function discord({ chat }) {
+  const msg = `join the discord for clips, vc, etc.: ${DISCORD}`
+  return msg
+}
