+++ commands/fc.js
@@ -1,0 +1,6 @@
+const { GAME_ID } = require('../vars')
+
+module.exports = function fc({ chat }) {
+  const msg = GAME_ID
+  return msg
+}
