+++ commands/hype.js
@@ -1,0 +1,4 @@
+module.exports = function hype({ chat }) {
+  const msg = 'yeeeeeaaaaaa boiiiiiii let\'s crack this rank meter'
+  return msg
+}
