+++ commands/lurk.js
@@ -1,0 +1,4 @@
+module.exports = function lurk({ chat }) {
+  const msg = 'thx for lurkin my dude'
+  return msg
+}
