+++ helpers/bitRegExp.js
@@ -1,0 +1,4 @@
+const bitPrefixes = ['xyz', 'abc']
+const bitRegExp = new RegExp(`((${bitPrefixes.join('|')})(\\d+))`, 'g')
+
+module.exports = bitRegExp
